Fix codegen bug where x10::runtime::Runtime was not found. Fix interaction with standalone x10rt implementation (avoid serialization if running with 1 place). Fix compiler warning in deserialization_dispatcher.
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@11328 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src-cpp/x10aux/deserialization_dispatcher.h
@@ -35,7 +35,7 @@ namespace x10aux {
         Deserializer *deser_v;
         size_t deser_sz;
 
-        int next_id;
+        size_t next_id;
 
         public:
         DeserializationDispatcher () : put_bfinder_v(NULL), put_bfinder_sz(0),
